restructure files, exclude prometheus and traefik
