--- server/prometheus-conf/prometheus.yml
@@ -1,14 +1,0 @@
-global:
-  scrape_interval: 20s
-  scrape_timeout: 15s
-  evaluation_interval: 20s
-
-scrape_configs:
-  - job_name: 'users'
-    scrape_interval: 15s
-    static_configs:
-    - targets: ["host.docker.internal:8081"]
-  - job_name: 'push_gateway'
-    scrape_interval: 15s
-    static_configs:
-    - targets: ['host.docker.internal:9090']
