--- server/prometheus-microservice/Makefile
@@ -1,29 +1,0 @@
-.PHONY: all
-all: build
-FORCE: ;
-
-SHELL  := env LIBRARY_ENV=$(LIBRARY_ENV) $(SHELL)
-LIBRARY_ENV ?= dev
-
-BIN_DIR = $(PWD)/bin
-
-.PHONY: build
-
-clean:
-	rm -rf bin/*
-
-dependencies:
-	go mod download
-
-build: dependencies build-prometheus
-
-build-prometheus:
-	go build -o $(path)prometheus cmd/app/main.go
-
-ci: dependencies test
-
-test:
-	go test -tags testing ./...
-
-fmt: ## gofmt and goimports all go files
-	find . -name '*.go' -not -wholename './vendor/*' | while read -r file; do gofmt -w -s "$$file"; goimports -w "$$file"; done
