--- server/prometheus-microservice/conf/conf.ini
@@ -1,5 +1,0 @@
-[server]
-Address = 8081
-ReadTimeout = 60
-WriteTimeout = 60
-MaxHeaderBytes = 1 << 20
