--- server/prometheus-microservice/go.mod
@@ -1,8 +1,0 @@
-module github.com/artiver/server/prometheus
-
-go 1.16
-
-require (
-	github.com/prometheus/client_golang v1.11.0 // indirect
-	gopkg.in/ini.v1 v1.62.0 // indirect
-)
