--- server/prometheus-microservice/run-prod.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-make build-prometheus path="./bin/"
-
-./bin/prometheus
